Rename bucket_distance, and other minor changes.
The name `bucket_distance` is misleading, as it gets larger for closer
values and is 512 for equal values. Let's deprecate that method and
replace it with `bucket_index`. This is also consistent with how it is
being used in kademlia_routing_table.

I added a comparison method that returns an `Ordering`, also because of
kademlia_routing_table, where this can replace some expressions where we
currently explicitly return `Ordering`s.

Finally, I did some minor cleanups and expanded the documentation.